added test for controller

+++ API/models/history.go
@@ -1,11 +1,11 @@
 package models
 
 type History struct {
-	ID              int
-	Username        string
-	FoodPurchase	  string
-	Price	          float
-	DeliveryMode	  string
-	Distance        float
-	CaloriesBurned	int
+	ID             int
+	Username       string
+	FoodPurchase   string
+	Price          float64
+	DeliveryMode   string
+	Distance       float64
+	CaloriesBurned int
 }
